Add a `nannou_playback.rs` example for visualising data from streams
Visualises the RGBA feed on the left, the depth feed on the right and
overloays the skeleton joint vertices on the RGBA feed.

Can be run by specifying a recorded file as a command line argument.

Note that we currently ignore the z component of the joint vertex
positions, but you can still get an idea of the data coming in. If
anyone does want to add 3D in the future, I recomment checking out the
nannou `vk_teapot.rs` example as a reference for working with 3D and
setting up the world, view and projection matrices.